Move localization helpers to shared scope
