--- EssentialApp/EssentialAppTests/Helpers/FeedUIIntegrationTests+Localization.swift
@@ -1,26 +1,0 @@
-//
-// Created by Rodrigo Porto.
-// Copyright © 2025 PortoCode. All Rights Reserved.
-//
-
-import Foundation
-import XCTest
-import EssentialFeed
-
-extension FeedUIIntegrationTests {
-    private class DummyView: ResourceView {
-        func display(_ viewModel: Any) {}
-    }
-    
-    var loadError: String {
-        LoadResourcePresenter<Any, DummyView>.loadError
-    }
-    
-    var feedTitle: String {
-        FeedPresenter.title
-    }
-    
-    var commentsTitle: String {
-        ImageCommentsPresenter.title
-    }
-}
